Imported fixes for SMAP (wrong watermark value) and IOMAN (incorrect declaration for lseek) stuff, fixed GSM extension's "ok" button, added patch entry for SRW IMPACT LE, removed NBNS's Rules.make file in favour of the common one, and adjusted in-game FAT driver's parameters and strategy to improve seeking performance.

+++ modules/iopcore/cdvdman/ioman_add.h
@@ -41,7 +41,7 @@ typedef struct _iop_ext_device_ops {
 	int	(*close)(iop_file_t *);
 	int	(*read)(iop_file_t *, void *, int);
 	int	(*write)(iop_file_t *, void *, int);
-	int	(*lseek)(iop_file_t *, unsigned long, int);
+	int	(*lseek)(iop_file_t *, int, int);
 	int	(*ioctl)(iop_file_t *, unsigned long, void *);
 	int	(*remove)(iop_file_t *, const char *);
 	int	(*mkdir)(iop_file_t *, const char *);
